RavenDB-20224 do not require nodeTag when /databases/*/indexes?namesOnly=true
